py/requirements.txt
fastapi
uvicorn[standard]
Pillow
requests
bs4
boto3
psycopg2-binary
python-dotenv
numpy
umap-learn
scikit-learn
google-genai
asyncpg
aiobotocore
deemix
deezer-py
google-auth
pysocks
pydantic-settings
ruff
